requirements.txt
moviepy
Pillow==9.5.0
# google-cloud-texttospeech
secure-smtplib
pyyaml